refactor: 공지 유형 색상 SoT 통합 + WARNING 색상 통일
- shared/config/announcement.ts 신설: 라벨/색상/헬퍼 단일 응집
- 어드민 배지 WARNING red-400 → red-500 보정 (배너/미리보기 기준)
- AnnouncementBanner TYPE_COLORS 인라인 제거, border 클래스만 컴포넌트에 잔존
- features/admin/labels.ts에서 ANNOUNCEMENT_TYPE_* 이동 (어드민 종속 아닌 시스템 enum)
